backend/requirements.txt
fastapi
uvicorn
langchain
langchain-groq
langchain-community
pypdf
python-multipart
python-dotenv